drivers/soc: icnss: Don't create debugfs when !CONFIG_ICNSS_DEBUG
The debugfs related codes are guarded using CONFIG_ICNSS_DEBUG
macro. Below that, there is another chunk of codes for debugfs.
If we disable CONFIG_DEBUG_FS, it will still execute icnss_debugfs_create()
because of the else-condition of the ifdeffery. Therefore, remove the
irrelevant function and make ICNSS_DEBUG depend on DEBUG_FS.
Guard the debugfs bits accordingly.

Change-Id: I53fab357ca99701ff5d33afdcea5c72b34c7b151
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: hridaya <hridaya@pixelos.net>